Refactor "Boost volume (8580)" to apply to any SID model
Since the libsidplayfp 2.5.0, the original boost option limited to 8580 is somewhat redundant, but the overall volume is still low which is a bigger issue, so it makes sense to expand the option to cover all SID models. The option is now "Boost volume" and has been moved from the "Emulation" section to the "Audio Output" section in the Preferences.