Reset SID states on emulation config change (edge cases & crash fix)
Edge case is when the SID emulation config is reset (e.g., toggling the DigiBoost in Preferences) which then fails to restore any toggled-off voices & filters (but in UI they are shown as restored falsely). If the Instant Seeking is enabled in such situation, you'd also get a crash due to prerender retaining an invalid reference to the SID decoder.