Make the "Check for updates" work on Linux
The linker issue was due to incompletely compiled wxWidgets (missing libcurl-dev). However, the wxWebRequest canceling is not usable reliably on Linux, so a workaround is introduced where any outdated response is ignored (rather than canceling the previous request immediately, which would cause segfaults due to opaque lifecycle of the wxWebRequest).