Generate the stil.index in ~/.config/username/sidplaywx/ on Linux
This is to support running from the AppImage and similar read-only containers. MSW however still uses the exe path (portable). Unlike for MSW builds, distributing pre-generated stil.index is moot on Linux -- not only because of this change, but also it seems that on Linux generating the stil.index doesn't slow down the initial app start.