Refactor Media keys support -- this feature is available only on MSW
The original method of polling media keys with a timer was made so it would be cross-compatible because the wxWindow::RegisterHotKey is implemented only on MSW. However it turns out that the wxGetKeyStateGtk doesn't support media keys at all (at least on Wayland?). Therefore I've enabled the "Media keys" option on Windows only, and reworked the media keys listener to use the more efficient native wxWindow::RegisterHotkey method.